Massive overhaul fixes
To get things out of the way first. I'm fixing the links from the "Bio", "Works", & "Projects" page, properly linked them to the files, and made sure that the files are working as intended. Also color corrected the nav bar so that its easier to see the text against the background.

Fixed the pictures not linking to the different html's/javascripts.

I'm still trying to fix the .css regarding the picture positioning.